Added field for editing title. Changing the title also changes the label for the region.
Wrapped the call to create_region() that was in
('.new_region_button').click inside of a function, because when passed
directly to click it causes an event argument to passed in as id, which
causes problems.

Changed create_region to expect id to be an integer or not set at all.